Documentation on how to make sortable tables work
The documentation seems to imply that sortable tables work out of the box.
When I added the relevant class names (sortable on table, sorted on th) it didn't work.
A bit of googling gave me this SO question and subsequent answer which worked. http://stackoverflow.com/a/26595432/365218

This is to include the proper steps needed to make sortable tables work.