Replace "header" by "heading" in the page content
Hi,

I do not mean to be picky but I think the term **"header" should be replaced by "heading"**, which is the most appropriate one to refer to the `<h1>`,` < h2>`, `<h3>`, `<h4>`, `<h5>`, `<h6>` html elements. You can have a look at it in:
* [WHATWG HTML - Living Standard](https://html.spec.whatwg.org/multipage/semantics.html#the-h1,-h2,-h3,-h4,-h5,-and-h6-elements)
* [W3C - HTML5 Standard](http://www.w3.org/TR/html5/sections.html#headings-and-sections)
* [MDN](https://developer.mozilla.org/en-US/docs/Web/HTML/Element/Heading_Elements)
* [W3schools] (http://www.w3schools.com/html/html_headings.asp)

I think it is a good idea in order to avoid confusion. I believe that in web development nowadays header is commonly used to refer to any of the following:
* Top area of a webpage, where we usually find the logo and main navigation menu
* Html element `<header>`
* HTTP headers

I have only performed the replacement within the visible content.

Thanks!

Best regards,
--
Diego