Unify structure and clarify usage in README.md
After reading the README.md several times I still had no clue how to successfully run the Docs on my local machine, so I looked through the issues hoping to find something usage related, and found #177 and #242 clearing things up a little, but still not completely. Based on brianveltman's step-by-step and jlukic's reply on it in #177 I was finally able to run the Docs locally. I expanded the How to Use guide to a structure with more levels and refined parts that were not clear enough for me. My goal was to end up with a How to that I can follow almost blindly and end up with the running Docs.