Clarify that dimmer.settings.closable is boolean
The documentation simply states that `closable` has a default of `auto` but doesn't state what other values it can take; I had to [check the code](https://github.com/Semantic-Org/Semantic-UI/blob/master/src/definitions/modules/dimmer.js#L364) to see that it's actually a boolean with logic to account for an `auto` flag. This change updates the docs to reflect this more clearly.